Server/src/requirements.txt
Django==2.2.1
Pillow==6.0.0
django-imagekit==4.0.2
django-material==1.5.8
python-memcached==1.59
django-notifications-hq==1.5.0
djangorestframework==3.10.2
social-auth-app-django==3.1.0
social-auth-core==3.2.0